Move font feature values map to FontContext
