+++ Cargo.toml
@@ -217,14 +217,14 @@ rustls-pki-types = "1.14"
 rustls-platform-verifier = "0.7.0"
 sea-query = { version = "1.0.1", default-features = false, features = ["backend-sqlite", "derive"] }
 sea-query-rusqlite = "0.8.0"
-selectors = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
+selectors = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
 seq-macro = "0.3.6"
 serde = "1.0.228"
 serde_bytes = "0.11"
 serde_core = "1.0.226"
 serde_json = "1.0"
 serde_test = "1.0"
-servo_arc = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
+servo_arc = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
 sha1 = "0.11"
 sha2 = "0.11"
 sha3 = "0.12"
@@ -234,12 +234,12 @@ smallvec = { version = "1.15", features = ["serde", "union"] }
 speexdsp-resampler = "0.1.0"
 string_cache = "0.9"
 strum = { version = "0.28", features = ["derive"] }
-stylo = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
-stylo_atoms = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
-stylo_dom = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
-stylo_malloc_size_of = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
-stylo_static_prefs = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
-stylo_traits = { git = "https://github.com/servo/stylo", rev = "5173fcd4455b99cc42777f81715a1bd68ed5b32c" }
+stylo = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
+stylo_atoms = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
+stylo_dom = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
+stylo_malloc_size_of = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
+stylo_static_prefs = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
+stylo_traits = { git = "https://github.com/servo/stylo", rev = "d3de91cbac7bba38e159239b3c0a360783fce2ee" }
 surfman = { version = "0.13.0", features = ["chains"] }
 swapper = "0.1"
 syn = { version = "2", default-features = false, features = ["clone-impls", "derive", "parsing"] }
@@ -453,15 +453,15 @@ codegen-units = 1
 #
 # Or for Stylo:
 #
-[patch."https://github.com/servo/stylo"]
-selectors = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-servo_arc = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-stylo = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-stylo_atoms = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-stylo_dom = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-stylo_malloc_size_of = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-stylo_static_prefs = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
-stylo_traits = { git = "https://github.com/simonwuelker/stylo", branch = "font-variant-alternates" }
+# [patch."https://github.com/servo/stylo"]
+# selectors = { path = "../stylo/selectors" }
+# servo_arc = { path = "../stylo/servo_arc" }
+# stylo = { path = "../stylo/style" }
+# stylo_atoms = { path = "../stylo/stylo_atoms" }
+# stylo_dom = { path = "../stylo/stylo_dom" }
+# stylo_malloc_size_of = { path = "../stylo/malloc_size_of" }
+# stylo_static_prefs = { path = "../stylo/stylo_static_prefs" }
+# stylo_traits = { path = "../stylo/style_traits" }
 #
 # Or for another Git dependency:
 #
